backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-openai
langchain-pinecone
langchain-community
pinecone
pypdf
tiktoken
